backend/reset_db.py
import os
from dotenv import load_dotenv
load_dotenv()

from app.database import engine, Base
from app import models

def reset_db():
    print("Dropping all tables...")
    Base.metadata.drop_all(bind=engine)
    print("Creating all tables...")
    Base.metadata.create_all(bind=engine)
    print("Database reset complete.")

if __name__ == "__main__":
    reset_db()
