backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-dotenv==1.0.0
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
passlib[argon2]==1.7.4
argon2-cffi
python-jose[cryptography]==3.3.0
multipart==0.0.5